gradle version is updated, library versions are updated, usecases in tracklist/domain are moved to tracklist/domain/usecase compose library is added ViewModelProvider.Factory is deprecated, replaced to ViewModelProvider.AndroidViewModelFactory
