+++ app/src/main/java/ru/akulinina/musicplayer/category/presentation/TrackCategoryViewModelProvider.kt
@@ -1,5 +1,6 @@
 package ru.akulinina.musicplayer.category.presentation
 
+import android.app.Application
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import ru.akulinina.musicplayer.category.domain.AddCategoryUseCase
@@ -7,11 +8,12 @@ import ru.akulinina.musicplayer.category.domain.GetCategoriesUseCase
 
 class TrackCategoryViewModelProvider(
     val getCategoriesUseCase: GetCategoriesUseCase,
-    val addCategoryUseCase: AddCategoryUseCase
-) : ViewModelProvider.Factory {
+    val addCategoryUseCase: AddCategoryUseCase,
+    application: Application
+) : ViewModelProvider.AndroidViewModelFactory(application) {
 
     @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return TrackCategoryViewModel(getCategoriesUseCase, addCategoryUseCase) as T
     }
 }
