+++ app/src/main/java/ru/akulinina/musicplayer/tracklist/presentation/TrackListViewModelProvider.kt
@@ -1,13 +1,16 @@
 package ru.akulinina.musicplayer.tracklist.presentation
 
+import android.app.Application
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import ru.akulinina.musicplayer.tracklist.domain.GetTrackListUseCase
+import ru.akulinina.musicplayer.tracklist.domain.usecase.GetTrackListUseCase
 
-class TrackListViewModelProvider(private val getTrackListUseCase: GetTrackListUseCase) : ViewModelProvider.Factory {
+class TrackListViewModelProvider(private val getTrackListUseCase: GetTrackListUseCase,
+                                 application: Application
+) : ViewModelProvider.AndroidViewModelFactory(application) {
 
     @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return TrackListViewModel(getTrackListUseCase) as T
     }
 }
